server/outreach/requirements.txt
Django
dj-database-url
psycopg2-binary
python-dotenv
pika
djangorestframework
langchain
langchain-google-genai
langchain-core
pydantic
chromadb
gunicorn
requests